requirements.txt
python-dotenv
requests
pydantic
fastapi
uvicorn
python-jose[cryptography]
python-multipart